Real_Time/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
EasyProcess==1.1
entrypoint2==1.1
filelock==3.9.0
idna==3.4
jeepney==0.8.0
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
mpmath==1.2.1
mss==9.0.1
networkx==3.0
numpy==1.24.3
Pillow==9.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
python3-xlib==0.15
pytweening==1.0.7
requests==2.28.1
scipy==1.10.1
sympy==1.11.1
torch==2.0.1+cpu
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.4.0
urllib3==1.26.13
